Sign artifacts on release
To deploy to maven central the artifacts need to be signed. We are using
the gpg maven plugin for this.

The maven.asc.enc is the gpg key encrypted with the public key of travis.
This is the recommended way of keeping secrets for travis in github is
described here https://docs.travis-ci.com/user/encrypting-files/.

We also need to provide the javadoc to publish to maven repo.